refactor: tidy up file names and imports/exports
