+++ package.json
@@ -6,7 +6,7 @@
   "engines": {
     "node": ">=18.18.0"
   },
-  "exports": "./dist/index.js",
+  "exports": "./dist/main.js",
   "files": [
     "src",
     "dist",
