--- src/index.ts
@@ -1,4 +1,0 @@
-import { createSdkClient } from './macros.js';
-import { SERVICES } from './service.js';
-
-export const RotaCloud = createSdkClient(SERVICES);
