+++ src/main.ts
@@ -1,0 +1,6 @@
+import { createSdkClient } from './client-builder.js';
+import { SERVICES } from './service.js';
+
+export * from './interfaces/index.js';
+export * from './models/index.js';
+export const createRotaCloudClient = createSdkClient(SERVICES);
